pyproject.toml
[project]
name = "duco-control"
version = "0.1.0"
description = "Driver scripts for the Duco-controlled 6-DoF F/T sensor."
requires-python = ">=3.10"
dependencies = [
    "pyserial>=3.5",
]

[project.scripts]
read-ft-sensor = "read_ft_sensor:main"

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
py-modules = ["read_ft_sensor"]
